Added dependecy for iohooks and relocated files
