+++ package.json
@@ -8,9 +8,9 @@
     "email": "contact@dygma.com"
   },
   "homepage": "https://github.com/Dygmalab/Bazecor",
-  "engines":{
-    "node":">=16.0.0",
-    "npm":">=8.11.0"
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=8.11.0"
   },
   "scripts": {
     "clean": "rm -rf dist && rm -rf node_modules",
@@ -108,6 +108,7 @@
     "source-map-support": "^0.5.16",
     "styled-components": "^5.2.3",
     "sudo-prompt": "^9.2.1",
+    "uiohook-napi": "^1.2.0",
     "usb": "^2.3.1"
   },
   "devDependencies": {
